+++ src/main/kotlin/hu/bme/sch/cmsch/controller/admin/TokensByTypeController.kt
@@ -1,7 +1,7 @@
 package hu.bme.sch.cmsch.controller.admin
 
 import hu.bme.sch.cmsch.admin.OverviewBuilder
-import hu.bme.sch.cmsch.dao.TokenPropertyRepository
+import hu.bme.sch.cmsch.repository.TokenPropertyRepository
 import hu.bme.sch.cmsch.dto.virtual.TokenPropertyVirtualEntity
 import hu.bme.sch.cmsch.dto.virtual.TokenStatVirtualEntity
 import hu.bme.sch.cmsch.util.getUser
