+++ src/main/kotlin/hu/bme/sch/cmsch/controller/admin/TokensStatsController.kt
@@ -1,17 +1,14 @@
 package hu.bme.sch.cmsch.controller.admin
 
 import hu.bme.sch.cmsch.admin.OverviewBuilder
-import hu.bme.sch.cmsch.dao.TokenPropertyRepository
-import hu.bme.sch.cmsch.dto.virtual.TokenPropertyVirtualEntity
 import hu.bme.sch.cmsch.dto.virtual.TokenStatVirtualEntity
+import hu.bme.sch.cmsch.repository.TokenPropertyRepository
 import hu.bme.sch.cmsch.util.getUser
 import hu.bme.sch.cmsch.util.getUserOrNull
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Controller
 import org.springframework.ui.Model
 import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import javax.servlet.http.HttpServletRequest
 
@@ -50,20 +47,22 @@ class TokensStatsController(
 
     private fun fetchOverview(): List<TokenStatVirtualEntity> {
         return tokenPropertyRepository.findAll()
-                .filter { it.token?.type == "default" }
-                .groupBy { it.token?.id }
-                .map { it.value }
-                .filter { it.isNotEmpty() }
-                .map { submissions ->
-                    TokenStatVirtualEntity(
-                        id =    submissions[0].token?.id ?: 0,
-                        token = submissions[0].token?.title ?: "n/a",
-                        type =  submissions[0].token?.type ?: "n/a",
-                        count = submissions.count { it.ownerUser?.groupName != organizerGroupName }
-                    )
-                }
-                .filter { it.count > 0 }
-                .sortedBy { it.token }
+            .asSequence()
+            .filter { it.token?.type == "default" }
+            .groupBy { it.token?.id }
+            .map { it.value }
+            .filter { it.isNotEmpty() }
+            .map { submissions ->
+                TokenStatVirtualEntity(
+                    id =    submissions[0].token?.id ?: 0,
+                    token = submissions[0].token?.title ?: "n/a",
+                    type =  submissions[0].token?.type ?: "n/a",
+                    count = submissions.count { it.ownerUser?.groupName != organizerGroupName }
+                )
+            }
+            .filter { it.count > 0 }
+            .sortedBy { it.token }
+            .toList()
     }
 
 }
