+++ src/main/kotlin/hu/bme/sch/cmsch/controller/api/AddonPaymentController.kt
@@ -1,7 +1,7 @@
 package hu.bme.sch.cmsch.controller.api
 
 import com.fasterxml.jackson.annotation.JsonView
-import hu.bme.sch.cmsch.dao.SoldProductRepository
+import hu.bme.sch.cmsch.repository.SoldProductRepository
 import hu.bme.sch.cmsch.dto.DebtDto
 import hu.bme.sch.cmsch.dto.FullDetails
 import hu.bme.sch.cmsch.service.RealtimeConfigService
