+++ src/main/kotlin/hu/bme/sch/cmsch/dto/RiddleCategoryDto.kt
@@ -1,6 +1,7 @@
 package hu.bme.sch.cmsch.dto
 
 import com.fasterxml.jackson.annotation.JsonView
+import hu.bme.sch.cmsch.dto.Preview
 
 data class RiddleCategoryDto(
