+++ src/main/kotlin/hu/bme/sch/cmsch/dto/view/RiddleSubmissionView.kt
@@ -1,4 +1,4 @@
-package hu.bme.sch.cmsch.dto.view
+package hu.bme.sch.cmsch.riddle.view
 
 import com.fasterxml.jackson.annotation.JsonView
 import hu.bme.sch.cmsch.dto.FullDetails
