+++ src/main/kotlin/hu/bme/sch/cmsch/model/RiddleEntity.kt
@@ -5,6 +5,7 @@ import hu.bme.sch.cmsch.admin.*
 import hu.bme.sch.cmsch.dto.Edit
 import hu.bme.sch.cmsch.dto.FullDetails
 import hu.bme.sch.cmsch.dto.Preview
+import hu.bme.sch.cmsch.model.ManagedEntity
 import javax.persistence.*
 
 @Entity
