+++ src/main/kotlin/hu/bme/sch/cmsch/repository/GroupToUserMappingRepository.kt
@@ -1,0 +1,12 @@
+package hu.bme.sch.cmsch.repository
+
+import hu.bme.sch.cmsch.model.GroupToUserMappingEntity
+import org.springframework.data.repository.CrudRepository
+import org.springframework.stereotype.Repository
+import java.util.*
+
+@Repository
+interface GroupToUserMappingRepository : CrudRepository<GroupToUserMappingEntity, Int> {
+    fun findByNeptun(neptun: String): Optional<GroupToUserMappingEntity>
+    fun findAllByGroupName(groupName: String): List<GroupToUserMappingEntity>
+}
