+++ src/main/kotlin/hu/bme/sch/cmsch/repository/GuildToUserMappingRepository.kt
@@ -1,0 +1,11 @@
+package hu.bme.sch.cmsch.repository
+
+import hu.bme.sch.cmsch.model.GuildToUserMappingEntity
+import org.springframework.data.repository.CrudRepository
+import org.springframework.stereotype.Repository
+import java.util.*
+
+@Repository
+interface GuildToUserMappingRepository : CrudRepository<GuildToUserMappingEntity, Int> {
+    fun findByNeptun(neptun: String): Optional<GuildToUserMappingEntity>
+}
