+++ src/main/kotlin/hu/bme/sch/cmsch/repository/LocationRepository.kt
@@ -1,0 +1,12 @@
+package hu.bme.sch.cmsch.repository
+
+import hu.bme.sch.cmsch.model.LocationEntity
+import org.springframework.data.repository.CrudRepository
+import org.springframework.stereotype.Repository
+import java.util.*
+
+@Repository
+interface LocationRepository : CrudRepository<LocationEntity, Int> {
+    fun findByUserId(userId: Int): Optional<LocationEntity>
+    override fun findAll(): List<LocationEntity>
+}
